rajout endpoint bestrating et delete, mise en place du calcul averagerating instantané, gestion erreur de multer, creation controllers suppression, rating, et bestrating

+++ Backend/controllers/books.js
@@ -5,12 +5,12 @@ exports.getAllBooks = async (req, res, next) => {
     try{
         const books = await Books.find();
         if(books.length === 0){
-            res.status(400).json({ message: 'Aucuns livres trouvés' });
+            return res.status(404).json({ message: 'Aucuns livres trouvés' });
         } else {            
-            res.status(200).json(books);
+            return res.status(200).json(books);
         }
     } catch (error) {
-        res.status(500).json({ error });
+        return res.status(500).json({ error });
     }
 };
 
@@ -18,12 +18,12 @@ exports.getOneBook = async (req, res, next) => {
     try {
         const book = await Books.findOne({ _id: req.params.id });
         if (!book) {
-            res.status(404).json({ message: 'Aucun livre trouvé' });
+            return res.status(404).json({ message: 'Aucun livre trouvé' });
         } else {
-            res.status(200).json(book);
+            return res.status(200).json(book);
         }
     } catch (error) {
-        res.status(500).json({ error });
+        return res.status(500).json({ error });
     }
 };
 
@@ -38,12 +38,12 @@ exports.createBook = async (req, res, next) => {
             imageUrl: `${req.protocol}://${req.get('host')}/images/${req.file.filename}`
         });
         await book.save();
-        res.status(201).json({ message: 'Livre créé !' });
+        return res.status(201).json({ message: 'Livre créé !' });
     } catch (error) {
         if (error.name === 'ValidationError') {
-            res.status(400).json({ error });
+            return res.status(400).json({ error });
         } else {
-            res.status(500).json({ error });
+            return res.status(500).json({ error });
         }
     }
 };
@@ -54,7 +54,7 @@ exports.modifyBook = async (req, res, next) => {
         const fileName = book.imageUrl.split('/images/')[1]; // récupération du nom du fichier image
         // verification que l'utilisateur est bien le propriétaire du livre
         if(req.auth.userId !== book.userId){
-            res.status(401).json({ message: 'Utilisateur non autorisé' });
+            return res.status(401).json({ message: 'Utilisateur non autorisé' });
         } else {
             delete req.body.userId; // suppression de l'id utilisateur envoyé par le client pour utiliser celui du token
             const bookObject = req.file ? // si une image est envoyée
@@ -65,11 +65,13 @@ exports.modifyBook = async (req, res, next) => {
                 } : { ...req.body, userId: req.auth.userId }; // si aucune image n'est envoyée
             
                 await Books.updateOne({ _id: req.params.id }, { ...bookObject, _id: req.params.id });
-                await fs.unlink(`images/${fileName}`); // suppression de l'ancienne image
-                res.status(200).json({ message: 'Livre modifié !' });
+                if(req.file){
+                    await fs.unlink(`images/${fileName}`); // suppression de l'ancienne image
+                }
+                return res.status(200).json({ message: 'Livre modifié !' });
         }
     } catch (error) {
-        res.status(500).json({ error });
+        return res.status(500).json({ error });
     }
 };
 
@@ -82,15 +84,15 @@ exports.deleteBook = async (req, res, next) => {
             const fileName = book.imageUrl.split('/images/')[1]; // récupération du nom du fichier image
             // verification que l'utilisateur est bien le propriétaire du livre 
             if(req.auth.userId !== book.userId){
-                res.status(401).json({ message: 'Utilisateur non autorisé !' });
+                return res.status(401).json({ message: 'Utilisateur non autorisé !' });
             } else {
                 await Books.deleteOne({ _id: req.params.id }); // suppression du livre
                 await fs.unlink(`images/${fileName}`); // suppression de l'image
-                res.status(200).json({ message: 'Livre supprimé !' });
+                return res.status(200).json({ message: 'Livre supprimé !' });
             }
         }
     } catch (error) {
-        res.status(500).json({ error });
+        return res.status(500).json({ error });
     }
 }
 
@@ -98,25 +100,38 @@ exports.ratingBook = async (req, res, next) => {
     try {
         const book = await Books.findOne({ _id: req.params.id });
         if (!book) {
-            res.status(404).json({ message: 'Aucun livre trouvé' });
-        } else {
-            const userRating = book.ratings.find(rating => rating.userId === req.auth.userId);
-            if (userRating) {
-                res.status(400).json({ message: 'Note déjà attribuée !' });
-            } else {
-                if(req.body.rating < 0 || req.body.rating > 5){
-                    res.status(400).json({ message: 'Note invalide !' });
-                }else{
-                    book.ratings.push({ userId: req.auth.userId, grade: req.body.rating });
-                    const allRatings = [...book.ratings, { userId: req.auth.userId, grade: req.body.rating }];
-                    const newAverageRating = Math.round(allRatings.reduce((sum, rating) => sum + rating.grade, 0) / allRatings.length);
-                    console.log(newAverageRating);
-                    await Books.updateOne({ _id: req.params.id }, { ratings: book.ratings, averageRating: newAverageRating});
-                    res.status(200).json({ message: 'Note attribuée !' });
-                }
-            }
+            return res.status(404).json({ message: 'Aucun livre trouvé' });
+        }
+        const userRating = book.ratings.find(rating => rating.userId === req.auth.userId);
+        if (userRating) {
+            return res.status(400).json({ message: 'Note déjà attribuée !' });
+        }
+        if(req.body.rating < 0 || req.body.rating > 5){
+            return res.status(400).json({ message: 'Note invalide !' });
+        }
+        book.ratings.push({ userId: req.auth.userId, grade: req.body.rating });
+        const newAverageRating = Math.round(book.ratings.reduce((sum, rating) => sum + rating.grade, 0) / book.ratings.length);
+        // récupération du livre mis à jour
+        const updatedBook = await Books.findOneAndUpdate(
+            { _id: req.params.id },
+            { $set: { ratings: book.ratings, averageRating: newAverageRating } },
+            { new: true, runValidators: true } 
+        );
+        return res.status(200).json(updatedBook); // Retourne le livre mis à jour
+    } catch (error) {
+        return res.status(500).json({ error });
+    }
+}
+
+exports.bestRatingBooks = async (req, res, next) => {
+    try {
+        const books = await Books.find().sort({ averageRating: -1 }).limit(3);
+        if(books.length === 0){
+            return res.status(404).json({ message: 'Aucuns livres trouvés' });
+        } else {            
+            return res.status(200).json(books);
         }
     } catch (error) {
-        res.status(500).json({ error });
+        return res.status(500).json({ error });
     }
 }
